refactor(providers): enable consumer groups with kafka provider
This commit attempts to enable consumer groups in the kafka
provider.

Unfortunately, due to the lack of proper support of the
Kafka provider in Rust, many hacks are necessary for this to work
properly we:

- Use `kafka-rs` for consumer groups support
- Use `rskafka` for topic creation/other control features

Without upstream commits to either `kafka-rs` or `rskafka` it is not
possible to accomplish the reasonable robustness goal of allowing
consumer groups without using *both* of these client libraries.

Another option is to use a library that depends on `rdkafka`, but that
introduces a shared library dependency that makes the provider a bit
more difficult to use than a pure Rust solution.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>