fix(tracing)!: return Dispatch from configure_tracing()
This commit returns Dispatch from `configure_tracing()` rather than
attempting to call `set_global_default()`[0].

`set_global_default` cannot safely be called from libraries
that are used in processes that will be executing (the host and a
binar provider, in the wasmCloud case).

Setting the default must be done from closer to the binar -- at the
very least not in a shared library that both the host and provider use.

[0]: https://docs.rs/tracing-core/0.1.32/tracing_core/dispatcher/fn.set_global_default.html

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>