feat(ci): Starts process of moving release pipelines away from Nix
This starts by adding a Dockerfile that can be used to build the image
with the given binary. I started by copying what we had in wadm and
adjusted from there. We can slowly migrate things over, but this
removes the image build step from the nix build for now

Signed-off-by: Taylor Thomas <taylor@oftaylor.com>